refactor(updater): migrate updater component to v2
Move from V1 structure (component.go + updaterimpl/) to V2 structure
(def/component.go, impl/, fx/fx.go) and update callers. Convert
fx.In/fx.Lifecycle dependencies to compdef.Lifecycle + Requires/Provides
pattern.
